Modularize reset_password and organize things
