+++ Guardfile
@@ -5,7 +5,7 @@ group :rspec do
     watch(%r[^lib/(.+)\.rb$])       { |m| "spec/#{m[1]}_spec.rb" }
     watch(%r[spec/spec_helper.rb])  { 'spec' }
     watch(%r[config])               { 'spec' }
-    watch(%r[^lib/account_manager/app.rb]) { 'spec/requests' }
+    watch(%r[^lib/account_manager/app.rb]) { 'spec' }
   end
 end
